docker

+++ install/sublime-text.sh
@@ -1,10 +1,13 @@
 #!/bin/bash
 
-wget -qO - https://download.sublimetext.com/sublimehq-pub.gpg | sudo apt-key add -
+if [ ! -e /etc/apt/sources.list.d/sublime-text.list ]; then
 
-sudo apt-get install -y apt-transport-https
+    wget -qO - https://download.sublimetext.com/sublimehq-pub.gpg | sudo apt-key add -
 
-echo "deb https://download.sublimetext.com/ apt/stable/" | sudo tee /etc/apt/sources.list.d/sublime-text.list
+    sudo apt-get install -y apt-transport-https
 
-sudo apt-get update -y
-sudo apt-get install -y sublime-text
+    echo "deb https://download.sublimetext.com/ apt/stable/" | sudo tee /etc/apt/sources.list.d/sublime-text.list
+
+    sudo apt-get update -y
+    sudo apt-get install -y sublime-text
+fi
